fix: fetch transcripts in lesson overview query
